pyproject.toml
[project]
name = "python-scrabble"
version = "0.1.0"
description = "Command-line Scrabble with AI opponents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama>=0.4.6",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.8.0",
    "pytest>=8.0.0",
    "types-colorama>=0.4.15",
]

[project.scripts]
scrabble = "game.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["game"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
packages = ["game"]

[[tool.mypy.overrides]]
module = "colorama.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "dotenv.*"
ignore_missing_imports = true
